+++ kinto/core/__init__.py
@@ -90,7 +90,7 @@ DEFAULT_SETTINGS = {
     'version_prefix_redirect_enabled': True,
     'trailing_slash_redirect_enabled': True,
     'multiauth.groupfinder': 'kinto.core.authorization.groupfinder',
-    'multiauth.policies': 'basicauth',
+    'multiauth.policies': '',
     'multiauth.policy.basicauth.use': ('kinto.core.authentication.'
                                        'BasicAuthAuthenticationPolicy'),
     'multiauth.authorization_policy': ('kinto.core.authorization.'