+++ kinto/plugins/accounts/__init__.py
@@ -3,8 +3,11 @@ import re
 from kinto.authorization import PERMISSIONS_INHERITANCE_TREE
 from pyramid.exceptions import ConfigurationError
 
-ACCOUNT_CACHE_KEY = 'accounts:{}:verified'
-ACCOUNT_POLICY_NAME = 'account'
+from .authentication import AccountsAuthenticationPolicy as AccountsPolicy
+from .utils import ACCOUNT_CACHE_KEY, ACCOUNT_POLICY_NAME
+
+
+__all__ = ['ACCOUNT_CACHE_KEY', 'ACCOUNT_POLICY_NAME', 'AccountsPolicy']
 
 DOCS_URL = "https://kinto.readthedocs.io/en/stable/api/1.x/accounts.html"
 
@@ -28,12 +31,12 @@ def includeme(config):
     settings = config.get_settings()
 
     # Check that the account policy is mentioned in config if included.
-    accountClass = 'AccountsAuthenticationPolicy'
+    accountClass = 'AccountsPolicy'
     policy = None
     for k, v in settings.items():
         m = re.match('multiauth\.policy\.(.*)\.use', k)
         if m:
-            if v.endswith(accountClass):
+            if v.endswith(accountClass) or v.endswith("AccountsAuthenticationPolicy"):
                 policy = m.group(1)
 
     if not policy: